requirements.txt

viam-sdk==0.45.2
typing-extensions
pyinstaller~=6.10.0
